refactor(app): extract AccessEligibilityGuard for authenticated layouts
Move auth, onboarding, and workspace eligibility checks out of the main
app layout so the same guard can wrap other authenticated surfaces.
